fix: clippy
